[rename]: 카테고리 Dto 디렉토리명 수정
