--- src/main/java/scs/planus/dto/todo/CategoryGetResponseDto.java
@@ -1,22 +1,0 @@
-package scs.planus.dto.todo;
-
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import scs.planus.domain.Color;
-import scs.planus.domain.TodoCategory;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class CategoryGetResponseDto {
-    private Long id;
-    private String name;
-    private Color color;
-
-    public CategoryGetResponseDto(TodoCategory todoCategory) {
-        this.id = todoCategory.getId();
-        this.name = todoCategory.getName();
-        this.color = todoCategory.getColor();
-    }
-}
